Move axes drawing to a new parent class AbstractPlotter

+++ src/main/java/de/tudresden/inf/mci/brailleplot/layout/AbstractCanvas.java
@@ -50,17 +50,17 @@ public abstract class AbstractCanvas<T extends PrintableData> {
         // New approach using a box model:
 
         // Create a page box
-        int pageWidth = mFormat.getProperty("page.width").toInt();
-        int pageHeight = mFormat.getProperty("page.height").toInt();
+        double pageWidth = mFormat.getProperty("page.width").toDouble();
+        double pageHeight = mFormat.getProperty("page.height").toDouble();
         Rectangle pageBox = new Rectangle(0, 0, pageWidth, pageHeight);
         mLogger.trace("Determined page box: {}", pageBox);
 
         // Create a margin box
         mLogger.trace("Cropping edges by defined margins:");
-        int marginTop = mFormat.getProperty("margin.top").toInt();
-        int marginLeft = mFormat.getProperty("margin.left").toInt();
-        int marginBottom = mFormat.getProperty("margin.bottom").toInt();
-        int marginRight = mFormat.getProperty("margin.right").toInt();
+        double marginTop = mFormat.getProperty("margin.top").toDouble();
+        double marginLeft = mFormat.getProperty("margin.left").toDouble();
+        double marginBottom = mFormat.getProperty("margin.bottom").toDouble();
+        double marginRight = mFormat.getProperty("margin.right").toDouble();
         Rectangle marginBox = new Rectangle(pageBox);
         try {
             marginBox.removeFromTop(marginTop);
